tracetypes: Fix location seek to no longer throw cast class exceptions.
Reviewed-on: https://git.eclipse.org/r/102523

+++ tracetypes/org.eclipse.tracecompass.incubator.traceevent.core/src/org/eclipse/tracecompass/incubator/internal/traceevent/core/trace/TraceEventTrace.java
@@ -146,7 +146,7 @@ public class TraceEventTrace extends TmfTrace implements ITmfPersistentlyIndexab
 
     @Override
     public double getLocationRatio(ITmfLocation location) {
-        return (double) getCurrentLocation().getLocationInfo() / fFile.length();
+        return ((Long) getCurrentLocation().getLocationInfo()).doubleValue() / fFile.length();
     }
 
     @Override
